Add files via upload
Our primary goal as an analytics team is to identify the ideal target customers for our
heart health supplement product line. To accomplish this, we look into the risk factors for heart
disease, such as age, BMI, smoking habits, and sleeping patterns.
The primary audience for this data presentation would be the marketing team of our
organization who need the data-driven insights to create a successful marketing strategy for the
heart health supplement product line in order to boost sales and increase revenue.
Our approach emphasizes prevention over cure, appealing to those who wish to
proactively manage their heart health. Therefore, it is designed to reach individuals who are at
risk for developing heart disease but who have not yet been diagnosed. This strategy aligns with
our mission to promote well-being while also appealing to a large customer base in order to
increase sales. We recognize that it is not enough to simply reach a large number of people, we
must also reach the right individuals. As a result, this strategy helps us to create a significant
influence by improving the health outcomes of our targeted demographics, while achieving
advertising success.